+++ cmd/changetype_test.go
@@ -5,8 +5,8 @@ import (
 	"io/ioutil"
 	"testing"
 
-	"github.com/rcmachado/changelog/chg"
-	"github.com/rcmachado/changelog/parser"
+	"github.com/cucumber/changelog/chg"
+	"github.com/cucumber/changelog/parser"
 	"github.com/stretchr/testify/assert"
 )